update viz draw functions
