+++ src/vizualization/MapViz.tsx
@@ -4,6 +4,10 @@ import { MapVizProps } from '../types'
 import { CountyLevelTooltip, ZipCodeLevelTooltip } from '../components/Tooltips'
 import { MapView } from './MapView'
 import { countyLevelData } from '../data/data'
+import {
+  drawZipCodeLevelCircles,
+  drawCountyLevelCircles,
+} from './ZipCodeLevelCircles'
 
 function MapViz({
   mainContainer,
@@ -47,112 +51,6 @@ function MapViz({
   // Path generator
   const path = d3.geoPath()
 
-  // Tippy instance for county level
-  let tippyInstanceCountyLevel: any
-
-  // Draw circles for county level
-  function drawCountyLevelCircles(
-    circlesData: any,
-    g: any,
-    levelUpdate: boolean = false
-  ) {
-    if (!circlesData) return
-    g.selectAll('.circle').remove()
-
-    // Circle Radius Scale
-    const radiusScale = d3
-      .scaleLog()
-      .domain([
-        d3.min(circlesData, (d: any) => Number(d.aggregateValue)),
-        d3.max(circlesData, (d: any) => Number(d.aggregateValue)),
-      ])
-      .range([5, 20])
-
-    const colorScales = d3
-      .scaleLinear<string>()
-      .domain(d3.extent(circlesData, (d: any) => d?.deliveryPrc) as any)
-      .range(['#33E48E', '#004223'] as [string, string])
-    if (circlesData.length) {
-      g.selectAll('circle')
-        .data(circlesData.filter((d: any) => d.x !== 0 && d.y !== 0))
-        .join('circle')
-        .attr('class', 'circle')
-        .attr('cx', (d: any) => d.x)
-        .attr('cy', (d: any) => d.y)
-        .attr('r', (d: any) => radiusScale(d.aggregateValue))
-        .attr('fill', (d: any) => colorScales(d.deliveryPrc) || '#ccc')
-        .attr('stroke', '#fff')
-        .attr('stroke-width', 0.5)
-        .style('opacity', 0.5)
-        .style('cursor', 'pointer')
-        .on('click', (event: any, d: any) => {
-          if (levelUpdate) return
-          const zipCodeLevelData = data.filter((x) => x.county === d.county)
-          if (d.aggregateValue === 1) return
-          drawZipCodeLevelCircles(zipCodeLevelData, g)
-          const zipCodeData = countiesJson.features.filter(
-            (x: any) => x.id === d.countyId
-          )
-          zoomToCounty(event.target, zipCodeData[0])
-        })
-        .on('mouseover', function (event: any, d: any) {
-          d3.select(this).attr('stroke', '#000')
-          d3.select(this).style('opacity', 1)
-          if (tippyInstanceCountyLevel) {
-            tippyInstanceCountyLevel.destroy()
-          }
-          tippyInstanceCountyLevel = CountyLevelTooltip(event, d)
-        })
-        .on('mouseout', function (event: any, d: any) {
-          d3.select(this).attr('stroke', '#fff')
-          d3.select(this).style('opacity', 0.5)
-        })
-    }
-  }
-
-  let tippyInstanceZipCodeLevel: any
-  // Draw zip code level circles
-  function drawZipCodeLevelCircles(circlesData: any, g: any) {
-    if (!circlesData) return
-    g.selectAll('.circle').remove()
-
-    // Circle radius scale
-    const radiusScale = d3
-      .scaleLog()
-      .domain([
-        d3.min(circlesData, (d: any) => Number(d.value)),
-        d3.max(circlesData, (d: any) => Number(d.value)),
-      ])
-      .range([5, 25])
-
-    if (circlesData.length) {
-      g.selectAll('circle')
-        .data(circlesData.filter((d: any) => d.x !== 0 && d.y !== 0))
-        .join('circle')
-        .attr('class', 'circle')
-        .attr('cx', (d: any) => d.x)
-        .attr('cy', (d: any) => d.y)
-        .attr('r', (d: any) => radiusScale(d.value) / transform.k)
-        .attr('fill', (d: any) =>
-          d.status === 'In Transit' ? '#33E48E' : '#004223'
-        )
-        .attr('stroke', '#fff')
-        .attr('stroke-width', 0.5)
-        .style('opacity', 0.5)
-        .style('cursor', 'pointer')
-        .on('click', (event: any, d: any) => {
-          event.stopPropagation()
-        })
-        .on('mouseover', function (event: any, d: any) {
-          if (d.length === 0) return
-          if (tippyInstanceZipCodeLevel) {
-            tippyInstanceZipCodeLevel.destroy()
-          }
-          tippyInstanceZipCodeLevel = ZipCodeLevelTooltip(event, d)
-        })
-    }
-  }
-
   // Draw state counties on state path click
   function drawStateCounties(counties: [], g: any, path: any) {
     g.selectAll('.county').remove()
@@ -171,32 +69,30 @@ function MapViz({
   // Draw map, based on view
   function updateView(view: 'states' | 'counties') {
     if (view === 'states') {
-      MapView(
-        stateJson.features,
+      MapView(stateJson.features, g, clicked, view, data)
+      drawCountyLevelCircles(
+        [],
         g,
-        clicked,
-        view,
-        data
+        false,
+        data,
+        transform,
+        zoomToCounty,
+        countiesJson
       )
-      drawCountyLevelCircles([], g, false)
     } else if (view === 'counties') {
-      MapView(
-        stateJson.features,
+      MapView(stateJson.features, g, clicked, view, data)
+      drawCountyLevelCircles(
+        countyLevelData(null, data),
         g,
-        clicked,
-        view,
-        data
+        true,
+        data,
+        transform,
+        zoomToCounty,
+        countiesJson
       )
-      drawCountyLevelCircles(countyLevelData(null, data), g, true)
     } else if (view === 'zipcodes') {
-      MapView(
-        stateJson.features,
-        g,
-        clicked,
-        view,
-        data
-      )
-      drawZipCodeLevelCircles(data, g)
+      MapView(stateJson.features, g, clicked, view, data)
+      drawZipCodeLevelCircles(data, g, transform)
     }
   }
 
@@ -211,9 +107,25 @@ function MapViz({
       data
     )
     if (view !== 'states') {
-      drawCountyLevelCircles(data, g, false)
+      drawCountyLevelCircles(
+        data,
+        g,
+        false,
+        data,
+        transform,
+        zoomToCounty,
+        countiesJson
+      )
     } else {
-      drawCountyLevelCircles([], g, true)
+      drawCountyLevelCircles(
+        [],
+        g,
+        true,
+        data,
+        transform,
+        zoomToCounty,
+        countiesJson
+      )
     }
   }
 
@@ -259,7 +171,15 @@ function MapViz({
       (county: any) => county.id.slice(0, 2) === d.id
     )
     drawStateCounties(filteredCounties, g, path)
-    drawCountyLevelCircles(countyLevelData(d.id, data), g)
+    drawCountyLevelCircles(
+      countyLevelData(d.id, data),
+      g,
+      false,
+      data,
+      transform,
+      zoomToCounty,
+      countiesJson
+    )
   }
 
   // Zoom event
